# packages/backend/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sistem POS API",
    "version": "1.0.0",
    "description": "Dokumentasi REST API Sistem POS real-time multi-tenant.",
    "contact": {
      "name": "Sistem POS Team",
      "email": "support@sistempos.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Terjadi kesalahan"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/dashboard/overview": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Data overview KPI dashboard",
        "description": "Mengambil data KPI untuk dashboard overview",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Tanggal mulai (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "end_date",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Tanggal akhir (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Data KPI berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "totalSales": {
                          "type": "number"
                        },
                        "totalTransactions": {
                          "type": "number"
                        },
                        "totalCustomers": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}